chatterbot/requirements.txt
ChatterBot==1.0.5
ChatterBot_corpus==1.2.0
transformers==4.28.1
textblob==0.15.3
requests==2.28.1
beautifulsoup4==4.11.1
nltk==3.7
pytorch==1.13.0
